Rydde litt, renamer fil med stavefeil, renamer komponent for lesbarhet, fikser typer
