+++ src/view/dialog/henvendelseInput/HenvendelseInputBox.tsx
@@ -10,8 +10,8 @@ import { useKladdContext } from '../../KladdProvider';
 import { dataOrUndefined, useOppfolgingContext, useUserInfoContext, useViewContext } from '../../Provider';
 import { HandlingsType, sendtNyHenvendelse } from '../../ViewState';
 import useHenvendelseStartTekst from '../UseHenvendelseStartTekst';
-import styles from './DialogInputBox.module.less';
-import HenvendelseInput from './HendvendelseInput';
+import HenvendelseInput from './HenvendelseInput';
+import styles from './HenvendelseInputBox.module.less';
 
 const maxMeldingsLengde = 5000;
 
@@ -26,15 +26,15 @@ const validerMelding = (melding: string, resten: any, props: { startTekst?: stri
         return 'Du må endre på meldingen';
     }
 };
-export type DialogInputProps = { melding: string };
-export type Handler = SubmitHandler<DialogInputProps>;
+export type HenvendelseInputBoxProps = { melding: string };
+export type Handler = SubmitHandler<HenvendelseInputBoxProps>;
 
 interface Props {
     dialog: DialogData;
     kanSendeHenveldelse: boolean;
 }
 
-const DialogInputBox = (props: Props) => {
+const HenvendelseInputBox = (props: Props) => {
     const bruker = useUserInfoContext();
     const oppfolgingContext = useOppfolgingContext();
     const oppfolging = dataOrUndefined(oppfolgingContext);
@@ -56,7 +56,7 @@ const DialogInputBox = (props: Props) => {
         melding: !!kladd?.tekst ? kladd.tekst : startTekst
     };
 
-    const validator = useFormstate<DialogInputProps>({
+    const validator = useFormstate<HenvendelseInputBoxProps>({
         melding: validerMelding
     });
 
@@ -146,4 +146,4 @@ const DialogInputBox = (props: Props) => {
     );
 };
 
-export default DialogInputBox;
+export default HenvendelseInputBox;
